Refactor coordinator management into two components
See DIH-233
