+++ src/router.jsx
@@ -19,6 +19,7 @@ import DestinationVolunteers from './sections/admin/destinations/destination/vol
 import DestinationEmails from './sections/admin/destinations/destination/emailTemplates';
 import DestinationAddVolunteer from './sections/admin/destinations/destination/addVolunteer';
 import DestinationCoordinators from './sections/admin/destinations/destination/coordinators';
+import DestinationAddCoordinator from './sections/admin/destinations/destination/addCoordinator';
 import Destinations from './sections/admin/destinations';
 import DestinationsTable from './sections/admin/destinations/destinationsTable';
 import AddDestination from './sections/admin/destinations/addDestination';
@@ -75,11 +76,14 @@ export default(
                             path="addvolunteer"
                             component={DestinationAddVolunteer}
                         />
-                        <Route
-                            name="Manage coordinators"
-                            path="coordinators"
-                            component={DestinationCoordinators}
-                        />
+                        <Route name="Coordinators" path="coordinators">
+                            <IndexRoute component={DestinationCoordinators} />
+                            <Route
+                                name="Add coordinator"
+                                path="new"
+                                component={DestinationAddCoordinator}
+                            />
+                        </Route>
                     </Route>
                 </Route>
                 <Route name="Users" path="admin/users">
