+++ src/sections/admin/destinations/destination/coordinators.jsx
@@ -2,30 +2,10 @@ import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
 import moment from 'moment';
 
-import { USER_ROLES } from '../../../../constants';
-
-import AddCoordinatorForm from './addCoordinatorForm';
 import Table from '../../../../commons/table';
-import { list } from '../../../../actions/userActions';
-import { retrieve, update } from '../../../../actions/destinationActions';
-import { pushNotification } from '../../../../actions/notificationActions';
+import { retrieve } from '../../../../actions/destinationActions';
 
-const createHandlers = (dispatch) => (
-    {
-        list() {
-            return dispatch(list());
-        },
-        retrieve(id) {
-            return dispatch(retrieve(id));
-        },
-        update(destination) {
-            return dispatch(update(destination));
-        },
-        notification(message, level) {
-            return dispatch(pushNotification(message, level));
-        }
-    }
-);
+const createHandlers = (dispatch) => (id) => dispatch(retrieve(id));
 
 class Coordinators extends Component {
     constructor(props) {
@@ -34,21 +14,7 @@ class Coordinators extends Component {
     }
 
     componentDidMount() {
-        this.handlers.list();
-        this.handlers.retrieve(this.props.params.destinationId);
-    }
-
-    filterModerators(items) {
-        if (!items) return [];
-        return items.filter(value => (
-            value.role === USER_ROLES.MODERATOR || value.role === USER_ROLES.ADMIN
-        ))
-        .map(value => (
-            {
-                id: value.id,
-                name: `${value.firstname} ${value.lastname}`
-            }
-        ));
+        this.handlers(this.props.params.destinationId);
     }
 
     normalizeCoordinatorObjectsForTable(items) {
@@ -66,26 +32,7 @@ class Coordinators extends Component {
         ));
     }
 
-    handleSubmit(data) {
-        const payload = {
-            users: [data],
-            id: this.props.destination.id
-        };
-        payload.users[0].userId = parseInt(payload.users[0].userId, 10);
-
-        this.handlers.update(payload)
-        .then(response => {
-            const message = 'Coordinator added';
-            const { error } = response;
-            if (!error) {
-                this.handlers.notification(message, 'success');
-                this.handlers.retrieve(this.props.params.destinationId);
-            }
-        });
-    }
-
     render() {
-        const moderators = this.filterModerators(this.props.users);
         return (
             <div>
                 <Table
@@ -98,25 +45,19 @@ class Coordinators extends Component {
                     itemKey="id"
                     items={this.normalizeCoordinatorObjectsForTable(this.props.destination.users)}
                 />
-                <AddCoordinatorForm
-                    moderators={moderators}
-                    onSubmit={e => { this.handleSubmit(e); }}
-                />
             </div>
         );
     }
 }
 
 const mapStateToProps = store => ({
-    destination: store.destinationState.destination,
-    users: store.userState.users
+    destination: store.destinationState.destination
 });
 
 Coordinators.propTypes = {
     destination: PropTypes.object.isRequired,
     dispatch: PropTypes.func.isRequired,
-    params: PropTypes.object.isRequired,
-    users: React.PropTypes.array.isRequired
+    params: PropTypes.object.isRequired
 };
 
 export default connect(mapStateToProps)(Coordinators);
