+++ src/sections/admin/destinations/destination/index.jsx
@@ -21,8 +21,12 @@ class Destination extends Component {
                     uri: `/admin/destinations/${this.props.params.destinationId}/addvolunteer`
                 },
                 {
-                    name: 'Manage coordinators',
+                    name: 'Coordinators',
                     uri: `/admin/destinations/${this.props.params.destinationId}/coordinators`
+                },
+                {
+                    name: 'Add coordinator',
+                    uri: `/admin/destinations/${this.props.params.destinationId}/coordinators/new`
                 },
                 {
                     name: 'Edit destination',
